--- app/src/main/java/com/haejung/snapmark/presentation/snap/SnapMarkEdit.kt
@@ -1,3 +1,0 @@
-package com.haejung.snapmark.presentation.snap
-
-class SnapMarkEdit ()
