+++ app/src/main/java/com/haejung/snapmark/presentation/snap/snapedit/SnapEditMark.kt
@@ -1,0 +1,3 @@
+package com.haejung.snapmark.presentation.snap.snapedit
+
+class SnapEditMark ()
